Add validation
This is not a great implementation, we are taking parameters from the request
as strings where they would ideally be other types (e.g dueDate should be
LocalDate not string). There is a problem with the serializer errors if we
used the correct types.

E.g with dueDate, if we passed the incorrect date format the serializer would
error before we even got to validation, and there is no obvious way to turn
serialization errors into nice messages for the front end.